+++ R/progress-server.R
@@ -150,7 +150,6 @@ builtin_handler_cli <- list(
 builtin_handler_progressr <- list(
   add = function(bar, .envir) {
     steps <- if (is.na(bar$total)) 0 else bar$total
-    bar$progressr_last <- 0L
     bar$progressr_progressor <- asNamespace("progressr")$progressor(
       steps,
       auto_finish = FALSE,
@@ -161,18 +160,14 @@ builtin_handler_progressr <- list(
   },
 
   set = function(bar, .envir) {
-    amount <- bar$current - bar$progressr_last
-    bar$last <- bar$current
-    if (!is.null(bar$progressr_progressor) && amount > 0) {
-      bar$progressr_progressor(amount = amount)
+    if (!is.null(bar$progressr_progressor)) {
+      bar$progressr_progressor(step = bar$current)
     }
   },
 
   complete = function(bar, .envir, result) {
-    amount <- bar$current - bar$progressr_last
-    bar$last <- bar$current
-    if (!is.null(bar$progressr_progressor) && amount > 0) {
-      bar$progressr_progressor(amount = amount, type = "finish")
+    if (!is.null(bar$progressr_progressor)) {
+      bar$progressr_progressor(step = bar$current, type = "finish")
     }
   },